modifications for relu fusion

+++ cuBLASLt/LtSgemm/sample_cublasLt_LtSgemm.h
@@ -41,9 +41,9 @@ void LtSgemm(cublasLtHandle_t ltHandle,
              int n,
              int k,
              const float *alpha, /* host pointer */
-             const float *A,
+             const __half *A,
              int lda,
-             const float *B,
+             const __half *B,
              int ldb,
              const float *beta, /* host pointer */
              float *C,
